Refactor: Merge account/mod.rs and account/subaccount.rs
